# docs/swagger.yaml
basePath: /
definitions:
  api.ScanRequest:
    properties:
      ip_range:
        type: string
    type: object
  model.Device:
    properties:
      first_seen:
        type: string
      hostname:
        type: string
      id:
        type: string
      ip_address:
        type: string
      last_seen:
        type: string
      mac_address:
        type: string
      manufacturer:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API to initiate a network scan and fetch discovered devices
  title: Network Scanner API
  version: "1.0"
paths:
  /clear:
    delete:
      description: Deletes all devices from in-memory store
      responses:
        "200":
          description: Devices cleared
          schema:
            type: string
      summary: Clear all stored devices (DEV ONLY)
  /devices:
    get:
      description: Returns the list of scanned devices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Device'
            type: array
      summary: Get all discovered devices
  /devices/{id}:
    get:
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Device'
        "404":
          description: Not found
          schema:
            type: string
      summary: Get device by ID
  /devices/{id}/tags:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Tag to remove
        in: body
        name: input
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Tag removed
          schema:
            type: string
      summary: Remove a tag from a device
    post:
      consumes:
      - application/json
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Tag to add
        in: body
        name: input
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Tag added
          schema:
            type: string
      summary: Add a tag to a device
  /devices/search:
    get:
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Device'
            type: array
      summary: Search devices by IP, hostname, or tags
  /scan:
    post:
      consumes:
      - application/json
      description: Starts scanning the given CIDR range in the background
      parameters:
      - description: IP range to scan
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/api.ScanRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            type: string
      summary: Initiate a network scan
schemes:
- http
swagger: "2.0"